+++ audio/index.js
@@ -1,9 +1,7 @@
-import { init_audio_ui } from './ui.js';
-window.init_audio_ui = init_audio_ui;
+import Init from './ui.js';
+import Stage from './stage.js'
+import Track from './track.js'
 
-import { mixer } from './mixer.js';
-window.MIXER = mixer;
-
-import Library from './library.js';
-window.TRACK_LIBRARY = Library.track;
-window.STAGE_LIBRARY = Library.stage;
+window.init_audio_ui = Init;
+window.STAGE_LIBRARY = Stage.library;
+window.TRACK_LIBRARY = Track.library;
