+++ audio/library.js
@@ -1,6 +1,3 @@
-import { Track } from './track.js';
-import { Stage } from './stage.js';
-
 /**
  * Library creates a generic local storage backed library for arbitrary
  * objects. The object must have an assign function describes how deserialize
@@ -128,40 +125,4 @@ class Library extends EventTarget {
     }
 }
 
-/**
- * The track library singleton
- * @extends {Library<Track>}
- */
-class TrackLibrary extends Library {
-    constructor() {
-        super(new Track());
-    }
-
-    /**
-     *
-     * @param {string} name
-     * @param {string} src
-     * @returns {[string, Track]}
-     */
-    find(name, src) {
-        return [...this.map()].find(([_, track]) => track.name === name || track.src === src);
-    }
-}
-
-/**
- * The stage library singleton
- * @extends {Library<Stage>}
- */
-class StageLibrary extends Library {
-    constructor() {
-        super(new Stage);
-    }
-}
-
-const track = new TrackLibrary();
-const stage = new StageLibrary();
-
-export default {
-    track,
-    stage
-}
+export default Library
